Snakes and Ladders 2
Board is created as well as buttons leading up to board however the background image is meant to change from the image with the buttons to a themed background image but it turns into a gray background and i can't fix it. I want the background image to change from the starting one to another which happens in the function def Board() line 80. Line 83 to 85 should change the background image but when the code is run it goes blank